refector: Button UI 분리 및 전송중 전송완료 로직 구현
