+++ luv-dub Watch App/CircleLoadingView.swift
@@ -20,10 +20,22 @@ struct CircleLoadingView: View {
                 .frame(width: 138, height: 138)
                 .foregroundColor(Color(red: 0.98, green: 0.07, blue: 0.31))
                 .rotationEffect(Angle.degrees(-90))
-                .animation(Animation.linear(duration: 2.5).repeatForever(autoreverses: false))
+                .animation(Animation.linear(duration: 1.5).repeatForever(autoreverses: false))
+            Text("전송중")
+                .font(
+                    Font.custom("Apple SD Gothic Neo", size: 15)
+                        .weight(.semibold)
+                )
+                .kerning(0.015)
+                .multilineTextAlignment(.center)
+                .foregroundColor(.white)
+                .padding(1.5)
         }
         .onAppear {
             isLoading = true
+            DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
+                isLoading = false 
+            }
         }
     }
